1_customer_support_saas/requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
streamlit
